created and moved functions to input handling file
